rename enum and move package
