+++ fabric-convention-tags-v2/src/main/resources/fabric-convention-tags-api-v2.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "net.fabricmc.fabric.mixin.convention",
+  "package": "net.fabricmc.fabric.mixin.tag",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
     "TagKeyMixin"
